refactor duplicated code into MobileMagnetometerService
